feat: spit models into their own files
