+++ app_flask/api/routes/sessions.py
@@ -1,7 +1,7 @@
 from flask import session, request
 from flask_imp.security import api_login_check
 
-from app_flask.models.sets import SetLogs
+from app_flask.models.set_logs import SetLogs
 from app_flask.models.workouts import Workouts, WorkoutSessions
 from .. import bp
